Add mileage estimate for the Boat class
Mileage estimate for the boat type depends on the time, wave
height (a random number between 1 and 100), and the boat weight,
which is a member variable for every Boat object.

Light boats are classified as those <= 10000 lbs. Heavy boats are
classified as those > 10000 lbs. The wave height is considered
high if the randomly generated number is > 50, and low if <= 50.

The if else statements in the Boat class's mileage estimate function
simulate that 1) light boats are more likely to travel faster than
heavier boats if the wave height is low, and 2) light boats are more
likely to travel slower than heavier boats if the wave height is high.

Every boat, no matter the wave height or weight, begins with a mileage
of 0.05 * time. Then, depending on the wave height and weight, a
certain percentage of that mileage is added to the final mileage.

If the waves are small and the boat is light, 50%-100% of the mileage
is added because these conditions will allow the boat to go faster.
If the waves are big and the boat is light, 1%-50% of the mileage
is added because these are conditions that will make the boat go
slower. 1%-50% is always added to a heavy boat because it will not be
affected by the waves, and heavier boats are assumed to always go slower.