Change random nums, boat mileage, & Skateboard var
Random numbers were incorrectly generated. Correct method is to
use the following formula to get a number within the range low-high:
rand() % (high - low + 1) + low.

Also edited computation for boat mileage (in addition to the random
number generation for simulating the wave height). A percent of the
initial mileage is now added.

Added a member data variable as well as some accessor and set functions
for the Skateboard class. This variable is named platFormLength and has
a default value of 24 inches.